gen.py changes
