+++ caoslib/offline/generate.py
@@ -1,0 +1,2 @@
+def run_generator():
+    pass
