+++ gen.py
@@ -1,0 +1,3 @@
+from caoslib.saving_tests import save_test
+#save_test(data, ans)
+from random import randint
